requirements.txt
selenium==4.10.0
pytest==9.0.3
webdriver-manager==3.8.6
allure-pytest==2.13.5
requests==2.33.0
pyyaml==6.0.1